+++ CHANGELOG.md
@@ -9,9 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.12.3 / 5.67.3] - 2023-11-
 
+### Added
+- added template to add usefull exclusions to confidential boxes
+
 ### Fixed
 - FIXED SECURITY ISSUE ID-23 SeManageVolumePrivilege is now blocked, as it allowed to read MFT data (thanks Diversenok)
 - fixed Program launch when forcing prcesses into a confidential box [#3173](https://github.com/sandboxie-plus/Sandboxie/issues/3173)
+
 
 
 ## [1.12.2 / 5.67.2] - 2023-11-